Move ExitException to separate file
